Fix enumeration on recent Windows
At least some recent Windows 11 builds insist on requesting on
StringDescriptors marked for EN_US, not EN. Because of that, we were
stalling string descriptor requests which left the device with a
Yellow bang and windows claiming it doesn't recognize the device.

Fix this simply by providing both EN and EN_US string descriptors.
Tested under Linux and Windows 11.